third_party/blink/web_tests/platform/generic/fast/forms/range/range-disabled-on-input-expected.txt

Check if disabling on value change on mousedown works:
PASS range1.value is "1"
PASS document.activeElement.id is "text1"
Check if disabling on value change on mousemvoe works:
PASS range1.value is "1"
PASS document.activeElement.id is "text1"
PASS successfullyParsed is true

TEST COMPLETE

 
